test(alerts): split scheduler retrieval coverage
